+++ src/modules/dashboard/components/ColonyHome/ColonyMembers/ColonyMembers.css
@@ -1,0 +1,65 @@
+.main {
+  display: block;
+  margin-top: 40px;
+}
+
+.main h4 {
+  display: inline-block;
+  font-size: var(--size-smallish);
+  color: var(--dark);
+  letter-spacing: var(--spacing-medium);
+}
+
+.main a h4:hover, .main a h4:active {
+  text-decoration: underline;
+  color: var(--colony-blue);
+}
+
+.userAvatars {
+  max-width: 130px;
+}
+
+.userAvatar {
+  display: inline-block;
+  margin: 0 15px 15px 0;
+  border: 2px solid transparent;
+  border-radius: 50px;
+}
+
+.userAvatar:nth-child(3n+0) {
+  margin-right: 0px;
+}
+
+.userAvatar:nth-child(13), .userAvatar:nth-child(14), .userAvatar:nth-child(15) {
+  margin-bottom: 0px;
+}
+
+.userAvatar:hover {
+  border-color: var(--primary);
+}
+
+.remaningAvatars {
+  display: inline-block;
+  margin: 0;
+  padding-top: 5px;
+  height: 30px;
+  width: 30px;
+  vertical-align: top;
+  border: 2px solid transparent;
+  border-radius: 50px;
+  background-color: var(--colony-white);
+  font-size: var(--size-tiny);
+  font-weight: var(--weight-bold);
+  text-align: center;
+  color: var(--temp-grey-13);
+  cursor: default;
+  letter-spacing: var(--spacing-medium);
+}
+
+.loadingText {
+  display: inline-block;
+  margin: -8px 0 0 5px;
+  vertical-align: middle;
+  font-size: var(--size-small);
+  letter-spacing: initial;
+}
