+++ src/modules/dashboard/components/ColonyHome/ColonyMembers/ColonyMembers.css.d.ts
@@ -1,0 +1,5 @@
+export const main: string;
+export const userAvatars: string;
+export const userAvatar: string;
+export const remaningAvatars: string;
+export const loadingText: string;
