+++ src/modules/dashboard/components/ColonyHome/ColonyMembers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyMembers';
