+++ src/modules/dashboard/components/ColonyMembers/ColonyMembers.css.d.ts
@@ -1,5 +1,5 @@
 export const main: string;
-export const userAvatars: string;
-export const userAvatar: string;
-export const remaningAvatars: string;
-export const loadingText: string;
+export const mainContentGrid: string;
+export const mainContent: string;
+export const rightAside: string;
+export const colonyList: string;
