+++ src/routes/routeConstants.ts
@@ -3,6 +3,8 @@ export const COLONY_HOME_ROUTE = '/colony/:colonyName';
 export const COLONY_EVENTS_ROUTE = `${COLONY_HOME_ROUTE}/events`;
 export const COLONY_EXTENSIONS_ROUTE = `${COLONY_HOME_ROUTE}/extensions`;
 export const PROGRAM_ROUTE = `${COLONY_HOME_ROUTE}/program/:programId`;
+export const MEMBERS_ROUTE = `${COLONY_HOME_ROUTE}/members`;
+export const MEMBERS_BY_DOMAIN_ROUTE = `${COLONY_HOME_ROUTE}/members/:domainId`;
 export const LEVEL_ROUTE = `${PROGRAM_ROUTE}/level/:levelId`;
 export const LEVEL_EDIT_ROUTE = `${PROGRAM_ROUTE}/level/:levelId/edit`;
 export const TASK_ROUTE = `${COLONY_HOME_ROUTE}/task/:draftId`;
